Initial work completed toward writing a .csv file importer.
This project is still in heavy development.  Too many changes to outline here.
This commit is of a functional .csv file reader that assumes the first three
columns are x,y,z positions and the fourth column is a magnitude value that gets
clipped on to the interval [0,1] and is assigned a color value.

Future work will move toward allowing the user to define which columns get to
be x,y,z and will incorporate all of the columns from the source .csv file
into the mesh somehow.